+++ CHANGELOG.md
@@ -1,12 +1,25 @@
 # Changelog
 
+## v3.1.0 - 2026-03-20
+
+### Features
+- Post event feedback form after event has ended to tracked event voice channels by @Spacesai1or in [#58](https://github.com/EvenLegion/Arbiter/pull/58)
+
+### Refactors
+- Event merit review menu now displays a single button per user, increasing the number of users displayed per page from 4 to 10 by @Spacesai1or in [#59](https://github.com/EvenLegion/Arbiter/pull/59)
+
+### Maintenance
+- Remove role select section in server Welcome message by @Spacesai1or in [#60](https://github.com/EvenLegion/Arbiter/pull/60)
+
 ## v3.0.0 - 2026-03-20
 
 ### Refactors
+
 - Complete overhaul of the codebase, optimizing for future contributors by @Spacesai1or in [#53](https://github.com/EvenLegion/Arbiter/pull/53)
 - Replace paid-for logging service with opensource self hosted option, allowing for more detailed logging without usage limits by @Spacesai1or in [#53](https://github.com/EvenLegion/Arbiter/pull/53)
 
 ### Maintenance
+
 - Init contributor's doc site explaining how the codebase works, how to get started, and how to release new versions of Arbiter by @Spacesai1or in [#53](https://github.com/EvenLegion/Arbiter/pull/53)
 
 ## v2.3.2 - 2026-03-15